backend/requirements.txt
easyocr==1.7.1
Flask==3.0.3
Flask_Cors==4.0.1
Flask_SocketIO==5.3.6
matplotlib==3.8.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pdf2image==1.17.0
Pillow==10.4.0
pytesseract==0.3.10
tensorflow==2.16.1
tensorflow_intel==2.16.1
Werkzeug==3.0.3
pyodbc==5.1.0
python-dotenv==1.0.1
watchdog==6.0.0